scripts/requirements.txt
anthropic
feedparser
python-slugify
requests
